requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
SQLAlchemy==2.0.32
protobuf~=4.25.4
ipython~=8.26.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
